requirements-x.txt
# X media upload (OAuth 1.0a) dependencies
requests-oauthlib>=1.3.1
